+++ History.md
@@ -1,5 +1,10 @@
 ## v.NEXT
 
+* `reactive-dict` now supports setting initial data when defining a named
+  `ReactiveDict`. No longer run migration logic when used on the server,
+  this is to prevent duplicate name error on reloads. Initial data is now
+  properly serialized.
+
 ## v1.4.4.2, 2017-05-02
 
 * Node has been upgraded to version 4.8.2.